+++ src/types.ts
@@ -4,3 +4,10 @@ export declare interface TokenInterface {
     expireAt: Date | null
     authState: object | null
 }
+
+export declare interface signInFunctionParams {
+    token: string,
+    tokenType: string | 'Bearer',
+    expiresIn: number,
+    authState: object
+}