+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 ## [1.0.3] - 
 ### Added
 - Documentation [https://authkit.arkadip.co](https://authkit.arkadip.co)
+- PrivateRoute now supports render and component props
+- Auth Token Type added on Sign In. Default `Bearer`
+
+### Changed
+- `Sign In` function Arguments is now an object
 
 ## [1.0.2] - 2020-08-29
 ### Added